added goals per game calculation

+++ db/migrate/004_create_players_table.rb
@@ -4,7 +4,7 @@ class CreatePlayersTable < ActiveRecord::Migration[4.2]
             t.string :name
             t.integer :series_goals
             t.integer :total_goals
-            t.integer :goals_per_game
+            t.float :goals_per_game
             t.integer :team_id
         end
     end
